==> Knowledge_Generation/2. BioBERT_encoding.ipynb <==
### Mount Google Drive

from google.colab import drive
drive.mount('/content/drive')

### Check the Drive Details

!ls "/content/drive/My Drive/COVID-19/Code_Base/Biobert_pretrained"

#### Install Libraries

!pip install biobert-embedding
!pip install pytorch_pretrained_bert

### Import Libraries

import glob
import pandas as pd
import nltk
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import AdaBoostClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.naive_bayes import MultinomialNB
from sklearn.neural_network import MLPClassifier
from sklearn.neighbors import KNeighborsClassifier 
from biobert_embedding.embedding import BiobertEmbedding

#### Read data set

df_BA_APIRO=pd.read_csv('/content/drive/My Drive/COVID-19/Code_Base/Data/Bert_Data/APIRO_Dataset.csv').drop(columns='Unnamed: 0',axis=1)
df_covid19_PICO=pd.read_csv('/content/drive/My Drive/COVID-19/Code_Base/Data/Bert_Data/df_covid19_PICO.csv')

#### BioBERT Model Initialization

bert_model_path='/content/drive/My Drive/COVID-19/Code_Base/Biobert_pretrained'
biobert = BiobertEmbedding(model_path=bert_model_path)

#### Biobert Encoding CB | Covid Data set



# CB Dataset
df_BA_APIRO['sent_length']=df_BA_APIRO['sent'].apply(lambda x: len(biobert.process_text(x))) 
df_BA_APIRO=df_BA_APIRO[df_BA_APIRO['sent_length']<=512]
# Covid 19 dataset
df_covid19_PICO['sent_length']=df_covid19_PICO['sent'].apply(lambda x: len(biobert.process_text(x))) 
df_covid19_PICO=df_covid19_PICO[df_covid19_PICO['sent_length']<=512]

def custom_sent_vector(cols):
    '''
    @Author - Fakhare Alam
    '''
    x=cols['sent']
    sent_vector=biobert.sentence_vector(x)
    return sent_vector

chunk_size = int(df_covid19_PICO.shape[0] / 100)
for start in range(0, df_BA_APIRO.shape[0], chunk_size):
    print('start set -' ,start)
    df_subset = df_BA_APIRO.iloc[start:start + chunk_size]
    df_subset['sent_embedding']=df_subset[['sent']].apply(custom_sent_vector,axis=1)
    df_subset.to_csv('/content/drive/My Drive/COVID-19/Code_Base/Data/Bert_Data/APIRO_Dataset_biobert_encoded_'+str(start)+'.csv',index=False)

chunk_size_CB = int(df_covid19_PICO.shape[0] / 10)
for start in range(0, df_covid19_PICO.shape[0], chunk_size_CB):
    print('start set -' ,start)
    df_subset = df_covid19_PICO.iloc[start:start + chunk_size_CB]
    df_subset['sent_embedding']=df_subset[['sent']].apply(custom_sent_vector,axis=1)
    df_subset.to_csv('/content/drive/My Drive/COVID-19/Code_Base/Data/Bert_Data/df_covid19_PICO_biobert_encoded_'+str(start)+'.csv',index=False)

#### Combine all Covid Files Together  CB | Covid

covid_path=r'/content/drive/My Drive/COVID-19/Code_Base/Data/Bert_Data/covid19'
all_files = glob.glob(covid_path + "/*.csv")
li = []
for filename in all_files:
    df = pd.read_csv(filename, index_col=None, header=0)
    li.append(df)

df_covid19_PICO_biobert_encoded_combined = pd.concat(li, axis=0, ignore_index=True)
df_covid19_PICO_biobert_encoded_combined.to_csv(covid_path+'/df_covid19_PICO_biobert_encoded_combined.csv',index=False)

df_covid19_PICO_biobert_encoded_combined.head()

CB_path=r'/content/drive/My Drive/COVID-19/Code_Base/Data/Bert_Data/CB'
all_files = glob.glob(CB_path + "/*.csv")
li = []
for filename in all_files:
    df = pd.read_csv(filename, index_col=None, header=0)
    li.append(df)

df_APIRO_Dataset_biobert_encoded_combined = pd.concat(li, axis=0, ignore_index=True)
df_APIRO_Dataset_biobert_encoded_combined.to_csv(CB_path+'/df_APIRO_Dataset_biobert_encoded_combined.csv',index=False)

df_APIRO_Dataset_biobert_encoded_combined.head()

df_APIRO_Dataset_biobert_encoded_combined.info()
